Work so far.  Which doesn't work.
I'm having a bugger of a time trying to figure out how to use TaskGroup with a perpetual feeder.  You can't use the TaskGroup from within any of its tasks, so you can't just have one of the tasks be the feeder.  This approach - merging the streams of network interface changes and results from the task group itself - should work, yet it doesn't; network changes are handled but probe tasks' results never emerge.