Added optional associated data for both `hit` and `miss` in Conclusion.
This looks like it'll be helpful in practice - e.g. maybe the probe attempt incidentally gathers information about the target host which is useful later.  It's better (safer and easier for users) to pass that data through the concurrency stack in a structured way, rather than expecting probe authors to figure out some side-band channel for it.

I also flipped the order of address & conclusion in the `description` of NetworkScanner.Result, to account for the conclusion now potentially having a bunch of data associated with it - much easier to read if that follows at the end of the description rather than at the start, ahead of the address.