Probe functions must now be Sendable.
This was required all along - the probe function is used concurrently from any number of Tasks - but the compiler wasn't detecting that or enforcing it.  Now, with the explicit annotations, it at least warns when you try to use a non-sendable function or closure.  (presumably in Swift 6 that'll become a hard error)