Now catch CancellationError and suppress it (from the iterator).
Apple doesn't appear to offer any guidance on whether one should throw this error out of your iterator, or just end iteration gracefully.  In the absence of that guidance, it doesn't make sense to me to throw an exception to the user of the iterator.  One way or another the caller caused the cancellation, so forcing them to suppress the exception seems redundant and annoying.

Note that this doesn't help in cases where the probe throws a different kind of cancellation error, e.g. URLError.cancelled.  Only the probe authors can handle that, since only they know what possible exceptions their probe can throw.