Removed redundant check for current task cancellation in `scan`.
Cancellation of the current task automatically triggers cancellation of the task group (per Swift's standard, unavoidale behaviour of propagating cancellation down the task hierarchy).  So if the current task is cancelled the task group must be too.

It's necessary to check the task group explicitly, however, because the task group can be cancelled independently, by one of its child tasks (a probe attempt) throwing an exception.  That cancellation won't automatically bubble up to the overarching task until the calling code consumes the failed task, ends iteration, *and* releases the iterator such that it is deallocated (so deinit can run and cancel the overarching task).  In most cases that will happen pretty quickly, but just in case it doesn't - e.g. the iterator user can't keep up with the results - it's better to cancel everything immediately.